backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic[email]>=2.5.0
openai>=1.12.0,<2.0.0
httpx>=0.27.0
scikit-learn==1.3.2
numpy>=1.24.0,<2.0.0
PyPDF2==3.0.1
python-docx==1.1.0
mangum==0.17.0
python-dotenv>=1.0.0
motor==3.3.2
pymongo==4.6.1
boto3==1.34.0
email-validator>=2.0.0